+++ app/Http/Controllers/Api/ResourceController.php
@@ -134,7 +134,7 @@ class ResourceController extends ApiController
                     'descript'          => isset($post['data']['description'])
                         ? $this->trans($post['data']['locale'], $post['data']['description'])
                         : null,
-                    'uri'               => Uuid::generate(4)->string,
+                    'uri'               => isset($post['data']['uri']) ? $post['data']['uri'] : Uuid::generate(4)->string,
                     'version'           => $post['data']['type'] == Resource::TYPE_AUTO ? 0 : 1,
                     'resource_type'     => $post['data']['type'],
                     'resource_url'      => isset($post['data']['resource_url']) ? $post['data']['resource_url'] : null,
@@ -165,6 +165,8 @@ class ResourceController extends ApiController
                     if (!empty($post['data']['created_at'])) {
                         $dbData['created_at'] = date('Y-m-d H:i:s', strtotime($post['data']['created_at']));
                     }
+
+                    $dbData['is_migrated'] = true;
                 }
 
                 $resource = Resource::create($dbData);
@@ -473,6 +475,13 @@ class ResourceController extends ApiController
                 if ($resource->is_reported == Resource::REPORTED_FALSE) {
                     Signal::where('resource_id', '=', $resource->id)->update(['status' => Signal::STATUS_PROCESSED]);
                 }
+            }
+
+            if (
+                isset($post['data']['migrated_data'])
+                && Auth::user()->username == 'migrate_data'
+            ){
+                $resource->is_migrated = true;
             }
 
             $resource->upl_freq_type = isset($post['data']['upl_freq_type']) ? $post['data']['upl_freq_type'] : null;