+++ app/Http/Controllers/DataController.php
@@ -597,10 +597,16 @@ class DataController extends Controller {
                     'dataset_uri' => $uri
                 ]
             ];
+            $resPerPage = 10;
+            $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+            $params['records_per_page'] = $resPerPage;
+            $params['page_number'] = $pageNumber;
+
             $rq = Request::create('/api/listResources', 'POST', $params);
             $apiResources = new ApiResource($rq);
             $res = $apiResources->listResources($rq)->getData();
             $resources = !empty($res->resources) ? $res->resources : [];
+            $resCount = $res->total_records;
 
             // get category details
             if (!empty($dataset->category_id)) {
@@ -731,6 +737,14 @@ class DataController extends Controller {
                 }
             }
 
+            $paginationData = $this->getPaginationData(
+                $resources,
+                $resCount,
+                array_except(app('request')->input(), ['rpage']),
+                $resPerPage,
+                'rpage'
+            );
+
             $dataset = $this->getModelUsernames($dataset);
             $discussion = $this->getForumDiscussion($dataset->forum_link);
 
@@ -740,10 +754,11 @@ class DataController extends Controller {
                 'user'          => $user,
                 'approved'      => (!empty($organisation) && $organisation->type == Organisation::TYPE_COUNTRY),
                 'dataset'       => $dataset,
-                'resources'     => $resources,
+                'resources'     => $paginationData['items'],
                 'buttons'       => $buttons,
                 'groups'        => $groups,
                 'setGroups'     => isset($setGroups) ? $setGroups : [],
+                'pagination'    => $paginationData['paginate']
             ];
 
             return view(
@@ -918,7 +933,8 @@ class DataController extends Controller {
                     $versionsPerPage
                 );
 
-                $resourcePaginationData = $this->getResourcePaginationData($data, $resource);
+                $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+                $resourcePaginationData = $this->getResourcePaginationData($data, $resource, $pageNumber);
 
                 return view(
                     'data/resourceView',