+++ app/Http/Controllers/UserController.php
@@ -568,6 +568,10 @@ class UserController extends Controller {
         // prepera request for resources
         unset($params['dataset_uri']);
         $params['criteria']['dataset_uri'] = $uri;
+        $resPerPage = 10;
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $params['records_per_page'] = $resPerPage;
+        $params['page_number'] = $pageNumber;
 
         $resourcesReq = Request::create('/api/listResources', 'POST', $params);
         $apiResources = new ApiResource($resourcesReq);
@@ -677,13 +681,22 @@ class UserController extends Controller {
             return redirect('/'. $root .'/datasets');
         }
 
+        $paginationData = $this->getPaginationData(
+            $resources->resources,
+            $resources->total_records,
+            array_except(app('request')->input(), ['rpage']),
+            $resPerPage,
+            'rpage'
+        );
+
         return view('user/datasetView', [
-            'class'     => 'user',
-            'groups'    => $groups,
-            'setGroups' => $setGroups,
-            'dataset'   => $this->getModelUsernames($dataset),
-            'resources' => $resources->resources,
-            'buttons'   => $buttons,
+            'class'      => 'user',
+            'groups'     => $groups,
+            'setGroups'  => $setGroups,
+            'dataset'    => $this->getModelUsernames($dataset),
+            'resources'  => $paginationData['items'],
+            'buttons'    => $buttons,
+            'pagination' => $paginationData['paginate']
         ]);
     }
 
@@ -778,6 +791,10 @@ class UserController extends Controller {
 
         $buttons['addResource'] = $rightCheck;
 
+        $resPerPage = 10;
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $params['records_per_page'] = $resPerPage;
+        $params['page_number'] = $pageNumber;
         $params['criteria']['dataset_uri'] = $uri;
 
         $resourcesReq = Request::create('/api/listResources', 'POST', $params);
@@ -906,6 +923,14 @@ class UserController extends Controller {
             return redirect('/'. $root .'/organisations/datasets/'. $organisation->uri);
         }
 
+        $paginationData = $this->getPaginationData(
+            $resources->resources,
+            $resources->total_records,
+            array_except(app('request')->input(), ['rpage']),
+            $resPerPage,
+            'rpage'
+        );
+
         return view(
             'user/orgDatasetView',
             [
@@ -913,10 +938,11 @@ class UserController extends Controller {
                 'groups'        => $groups,
                 'setGroups'     => $setGroups,
                 'dataset'       => $datasetData,
-                'resources'     => $resources->resources,
+                'resources'     => $paginationData['items'],
                 'activeMenu'    => 'organisation',
                 'fromOrg'       => $organisation,
                 'buttons'       => $buttons,
+                'pagination'    => $paginationData['paginate']
             ]
         );
     }
@@ -2236,7 +2262,8 @@ class UserController extends Controller {
             $request->session()->flash('alert-success', __('custom.delete_error'));
         }
 
-        $resourcePaginationData = $this->getResourcePaginationData($data, $resource);
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $resourcePaginationData = $this->getResourcePaginationData($data, $resource, $pageNumber);
 
         return view('user/resourceView', [
             'class'         => 'user',
@@ -2357,7 +2384,8 @@ class UserController extends Controller {
             $data = isset($resultConvert->data) ? $resultConvert->data : [];
         }
 
-        $resourcePaginationData = $this->getResourcePaginationData($data, $resource);
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $resourcePaginationData = $this->getResourcePaginationData($data, $resource, $pageNumber);
 
         // handle delete submit
         if ($request->has('delete')) {
@@ -5020,6 +5048,10 @@ class UserController extends Controller {
 
         $buttons[$datasetData->uri]['delete'] = $rightCheck;
 
+        $resPerPage = 10;
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $params['records_per_page'] = $resPerPage;
+        $params['page_number'] = $pageNumber;
         $params['criteria']['dataset_uri'] = $uri;
 
         $resourcesReq = Request::create('/api/listResources', 'POST', $params);
@@ -5148,6 +5180,14 @@ class UserController extends Controller {
             return redirect('/'. $root .'/groups/datasets/'. $group->uri);
         }
 
+        $paginationData = $this->getPaginationData(
+            $resources->resources,
+            $resources->total_records,
+            array_except(app('request')->input(), ['rpage']),
+            $resPerPage,
+            'rpage'
+        );
+
         return view(
             'user/groupDatasetView',
             [
@@ -5155,10 +5195,11 @@ class UserController extends Controller {
                 'groups'        => $groups,
                 'setGroups'     => $setGroups,
                 'dataset'       => $datasetData,
-                'resources'     => $resources->resources,
+                'resources'     => $paginationData['items'],
                 'activeMenu'    => 'group',
                 'buttons'       => $buttons,
                 'group'         => $group,
+                'pagination'    => $paginationData['paginate']
             ]
         );
     }
@@ -5448,7 +5489,8 @@ class UserController extends Controller {
             $request->session()->flash('alert-success', __('custom.delete_error'));
         }
 
-        $resourcePaginationData = $this->getResourcePaginationData($data, $resource);
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $resourcePaginationData = $this->getResourcePaginationData($data, $resource, $pageNumber);
 
         return view('user/groupResourceView', [
             'class'         => 'user',